+++ src/pyalab/steps/__init__.py
@@ -1,6 +1,6 @@
-from .base import AspirateParameters
-from .base import DispenseParameters
 from .base import Step
+from .params import AspirateParameters
+from .params import DispenseParameters
 from .set_volume import SetInitialVolume
 from .set_volume import SetVolume
 from .transfer import Transfer
