+++ src/pyalab/steps/params.py
@@ -1,0 +1,9 @@
+from pyalab.steps.base import LiquidMovementParameters
+
+
+class AspirateParameters(LiquidMovementParameters, frozen=True):
+    pass
+
+
+class DispenseParameters(LiquidMovementParameters, frozen=True):
+    pass
